// src/modules/edc-dmgmt-client/model/policy.ts
/**
 * EDC REST API
 * All files merged by open api merger
 *
 * The version of the OpenAPI document: 1.0.0-SNAPSHOT
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Prohibition } from './prohibition';
import { Duty } from './duty';
import { Permission } from './permission';


export interface Policy {
    uid: string;
    permissions?: Array<Permission>;
    prohibitions?: Array<Prohibition>;
    obligations?: Array<Duty>;
    extensibleProperties?: { [key: string]: object; };
    inheritsFrom?: string;
    assigner?: string;
    assignee?: string;
    target?: string;
    type: Policy.TypeEnum;
}
export namespace Policy {
    export type TypeEnum = 'SET' | 'OFFER' | 'CONTRACT';
    export const TypeEnum = {
        Set: 'SET' as TypeEnum,
        Offer: 'OFFER' as TypeEnum,
        Contract: 'CONTRACT' as TypeEnum
    };
}